Inventory_service/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
cloudinary
grpcio
grpcio-tools
pytz